🔨 Refactor(#120): Learn 컴포넌트 HOC 패턴 사용하여 데이터 분리
